Move JWT stuff into JwtTools module
JwtTools::JwtAuthStrategy - Warden authentication strategy using JWT tokens in Authorization header
JwtTools::JwtCoder - encoding/decoding of JWTs
